修改角色枚举类的数据类型；删除 PlatformUserMapper 中的多余方法；修改 PlatformUser 测试中的方法，是用 Wrapper 方式查询；修改用户登陆逻辑，允许多角色登录，并返回用户的所有角色；修改 SaToken 获取用户角色列表和获取用户授权资源的逻辑；修改修改用户补充角色信息的权限范围；优化用户登陆逻辑，减少数据库查询次数，平均市场从 1800ms 优化到 1300ms；

+++ src/main/java/com/scnujxjy/backendpoint/constant/enums/RoleEnum.java
@@ -14,11 +14,11 @@ public enum RoleEnum {
 
     XUELIJIAOYUBU_ADMIN(3, "学历教育部管理员");
 
-    private Integer roleId;
+    private Long roleId;
 
     private String roleName;
 
-    RoleEnum(int roleId, String roleName) {
+    RoleEnum(long roleId, String roleName) {
         this.roleId = roleId;
         this.roleName = roleName;
     }
